refactor tests to call reset directly
